app2/requirements.txt
flask
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
sendgrid
